Jvmconnection controller done

+++ Deadlock_detection_tool/src/main/java/com/harsh/DDT/dao/RemoteServerRepository.java
@@ -6,6 +6,6 @@ import com.harsh.DDT.model.RemoteServer;
 
 
 @Repository
-public interface RemoteServerRepository extends JpaRepository<RemoteServer,Integer> {
+public interface RemoteServerRepository extends JpaRepository<RemoteServer,Long> {
 
 }
